Attempt to fix #12
Not sure what would cause the @PROCS array to have undef values, I can
think of a few things. But the fact it happens in a DESTROY tells me it
is happening in cleanup, so the procs it did contain have probably
already been garbage collected turning them to undef entries in the
array.

Since this is likely a garbage collection order issue, the best solution
is this. This will skip over undef entries. Since it is cleanup and
everything is deconstructing anyway.